Project structure changed. Helper functions moved respectively to serviceplatformen_cpr/helpers/'. Tests passed after refactoring.
